+++ build_runner/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.8.4
 
 - Log the number of completed actions on successful builds.
+- Support the new `isRoot` field for `BuilderOptions` so that builders can
+  do different things for the root package.
 
 ## 0.8.3